move app reducers and rename cart reducers
